fixing a bunch of inconsistencies and potential bugs as indicated by findbugs, pmd and eclipse
git-svn-id: http://google-refine.googlecode.com/svn/trunk@2301 7d457c2a-affb-35e4-300a-418c747d4874

+++ main/src/com/google/refine/browsing/util/ExpressionTimeValueBinner.java
@@ -170,7 +170,7 @@ public class ExpressionTimeValueBinner implements RowVisitor, RecordVisitor {
                 long t = ((Date) value).getTime();
                 hasTime = true;
                     
-                int bin = (int) Math.floor((t - _index.getMin()) / _index.getStep());
+                int bin = (int) Math.floor((double) (t - _index.getMin()) / (double) _index.getStep());
                 if (bin >= 0 && bin < bins.length) { // as a precaution
                     bins[bin]++;
                 }
